requirements.txt
# Django
Django
pillow
djangorestframework==3.14.0
django-cors-headers==3.13.0